nvmf: widen subsystem state change ctx nsid to uint32_t
The nsid field of struct nvmf_subsystem_state_change_ctx was uint16_t,
while the entire pause API path uses uint32_t. Assigning the uint32_t
argument into the field truncated it.

Impact: SPDK_NVME_GLOBAL_NS_TAG (0xFFFFFFFF) was truncated to 0xFFFF, so
the "pause all namespaces" branch in nvmf_poll_group_pause_subsystem
(guarded by nsid == SPDK_NVME_GLOBAL_NS_TAG) was never taken for a global
pause routed through the context. Instead it fell into the single-ns
branch and paused no namespace at all, breaking subsystem-wide quiesce.

Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/29007
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>

+++ lib/nvmf/nvmf_internal.h
@@ -347,7 +347,7 @@ struct spdk_nvmf_ctrlr {
 
 struct nvmf_subsystem_state_change_ctx {
 	struct spdk_nvmf_subsystem			*subsystem;
-	uint16_t					nsid;
+	uint32_t					nsid;
 
 	enum spdk_nvmf_subsystem_state			original_state;
 	enum spdk_nvmf_subsystem_state			requested_state;
